requirements.txt
openpyxl==3.0.7
xlwt==1.3.0
xlrd==2.0.1
requests==2.20.0
pandas==2.0.2
PyYAML==6.0
jsonpath==0.82